[compdev-683] Extend the WPML integration to translate the field labels
The field labels can be group in a single package.
We could create a new package kind just for that, but it would be used
only for 1 package.
Instead, I decide to use a non-numerial custom user ID ("-- Labels --")
for this specific package, so we are sure it won't collide with a
real numerical user ID.

https://onthegosystems.myjetbrains.com/youtrack/issue/compdev-683/